move wfm tests and fix unnamed args in Variable
